Refactor: split queue IPC, owner runtime, and CLI internal paths
