+++ src/cli.ts
@@ -11,6 +11,8 @@ import {
   listBuiltInAgents,
   resolveAgentCommand as resolveAgentCommandFromRegistry,
 } from "./agent-registry.js";
+import { parseQueueOwnerFlags } from "./cli-internal-owner.js";
+import { configurePublicCli } from "./cli-public.js";
 import {
   initGlobalConfigFile,
   loadResolvedConfig,
@@ -49,7 +51,6 @@ import {
   EXIT_CODES,
   NON_INTERACTIVE_PERMISSION_POLICIES,
   OUTPUT_FORMATS,
-  PERMISSION_MODES,
   type NonInteractivePermissionPolicy,
   type AuthPolicy,
   type OutputFormat,
@@ -105,17 +106,6 @@ type StatusFlags = {
   session?: string;
 };
 
-type QueueOwnerFlags = {
-  sessionId: string;
-  ttlMs: number;
-  permissionMode: PermissionMode;
-  nonInteractivePermissions?: NonInteractivePermissionPolicy;
-  authPolicy?: AuthPolicy;
-  timeoutMs?: number;
-  verbose?: boolean;
-  suppressSdkConsoleErrors?: boolean;
-};
-
 const TOP_LEVEL_VERBS = new Set([
   "__queue-owner",
   "prompt",
@@ -168,31 +158,6 @@ function parseTimeoutSeconds(value: string): number {
     throw new InvalidArgumentError("Timeout must be a positive number of seconds");
   }
   return Math.round(parsed * 1000);
-}
-
-function parseTimeoutMilliseconds(value: string): number {
-  const parsed = Number(value);
-  if (!Number.isFinite(parsed) || parsed <= 0) {
-    throw new InvalidArgumentError("Timeout must be a positive number of milliseconds");
-  }
-  return Math.round(parsed);
-}
-
-function parseNonNegativeMilliseconds(value: string): number {
-  const parsed = Number(value);
-  if (!Number.isFinite(parsed) || parsed < 0) {
-    throw new InvalidArgumentError("TTL must be a non-negative number of milliseconds");
-  }
-  return Math.round(parsed);
-}
-
-function parsePermissionMode(value: string): PermissionMode {
-  if (!PERMISSION_MODES.includes(value as PermissionMode)) {
-    throw new InvalidArgumentError(
-      `Invalid permission mode "${value}". Expected one of: ${PERMISSION_MODES.join(", ")}`,
-    );
-  }
-  return value as PermissionMode;
 }
 
 export function parseTtlSeconds(value: string): number {
@@ -1828,108 +1793,6 @@ function detectJsonStrict(argv: string[]): boolean {
   return false;
 }
 
-function parseQueueOwnerFlags(argv: string[]): QueueOwnerFlags | undefined {
-  if (argv[0] !== "__queue-owner") {
-    return undefined;
-  }
-
-  const flags: Partial<QueueOwnerFlags> = {
-    ttlMs: DEFAULT_QUEUE_OWNER_TTL_MS,
-  };
-
-  const consumeValue = (
-    index: number,
-    token: string,
-  ): { value: string; next: number } => {
-    if (token.includes("=")) {
-      return {
-        value: token.slice(token.indexOf("=") + 1),
-        next: index,
-      };
-    }
-    const value = argv[index + 1];
-    if (!value || value.startsWith("-")) {
-      throw new InvalidArgumentError(`${token} requires a value`);
-    }
-    return {
-      value,
-      next: index + 1,
-    };
-  };
-
-  for (let index = 1; index < argv.length; index += 1) {
-    const token = argv[index];
-    if (token === "--session-id" || token.startsWith("--session-id=")) {
-      const consumed = consumeValue(index, token);
-      flags.sessionId = parseNonEmptyValue("Session id", consumed.value);
-      index = consumed.next;
-      continue;
-    }
-    if (token === "--ttl-ms" || token.startsWith("--ttl-ms=")) {
-      const consumed = consumeValue(index, token);
-      flags.ttlMs = parseNonNegativeMilliseconds(consumed.value);
-      index = consumed.next;
-      continue;
-    }
-    if (token === "--permission-mode" || token.startsWith("--permission-mode=")) {
-      const consumed = consumeValue(index, token);
-      flags.permissionMode = parsePermissionMode(consumed.value);
-      index = consumed.next;
-      continue;
-    }
-    if (
-      token === "--non-interactive-permissions" ||
-      token.startsWith("--non-interactive-permissions=")
-    ) {
-      const consumed = consumeValue(index, token);
-      flags.nonInteractivePermissions = parseNonInteractivePermissionPolicy(
-        consumed.value,
-      );
-      index = consumed.next;
-      continue;
-    }
-    if (token === "--auth-policy" || token.startsWith("--auth-policy=")) {
-      const consumed = consumeValue(index, token);
-      flags.authPolicy = parseAuthPolicy(consumed.value);
-      index = consumed.next;
-      continue;
-    }
-    if (token === "--timeout-ms" || token.startsWith("--timeout-ms=")) {
-      const consumed = consumeValue(index, token);
-      flags.timeoutMs = parseTimeoutMilliseconds(consumed.value);
-      index = consumed.next;
-      continue;
-    }
-    if (token === "--verbose") {
-      flags.verbose = true;
-      continue;
-    }
-    if (token === "--suppress-sdk-console-errors") {
-      flags.suppressSdkConsoleErrors = true;
-      continue;
-    }
-    throw new InvalidArgumentError(`Unknown __queue-owner option: ${token}`);
-  }
-
-  if (!flags.sessionId) {
-    throw new InvalidArgumentError("__queue-owner requires --session-id");
-  }
-  if (!flags.permissionMode) {
-    throw new InvalidArgumentError("__queue-owner requires --permission-mode");
-  }
-
-  return {
-    sessionId: flags.sessionId,
-    ttlMs: flags.ttlMs ?? DEFAULT_QUEUE_OWNER_TTL_MS,
-    permissionMode: flags.permissionMode,
-    nonInteractivePermissions: flags.nonInteractivePermissions,
-    authPolicy: flags.authPolicy,
-    timeoutMs: flags.timeoutMs,
-    verbose: flags.verbose,
-    suppressSdkConsoleErrors: flags.suppressSdkConsoleErrors,
-  };
-}
-
 function emitJsonErrorEvent(error: NormalizedOutputError): void {
   const formatter = createOutputFormatter("json", {
     jsonContext: {
@@ -1986,8 +1849,10 @@ export async function main(argv: string[] = process.argv): Promise<void> {
     requestedOutputFormat,
     requestedJsonStrict,
   );
-  const internalQueueOwnerFlags = parseQueueOwnerFlags(argv.slice(2));
-  const builtInAgents = listBuiltInAgents(config.agents);
+  const internalQueueOwnerFlags = parseQueueOwnerFlags(
+    argv.slice(2),
+    DEFAULT_QUEUE_OWNER_TTL_MS,
+  );
 
   const program = new Command();
   program
@@ -2008,65 +1873,20 @@ export async function main(argv: string[] = process.argv): Promise<void> {
   }
 
   addGlobalFlags(program);
-
-  for (const agentName of builtInAgents) {
-    registerAgentCommand(program, agentName, config);
-  }
-
-  registerDefaultCommands(program, config);
-
-  const scan = detectAgentToken(argv.slice(2));
-  if (
-    !scan.hasAgentOverride &&
-    scan.token &&
-    !TOP_LEVEL_VERBS.has(scan.token) &&
-    !builtInAgents.includes(scan.token)
-  ) {
-    registerAgentCommand(program, scan.token, config);
-  }
-
-  program.argument("[prompt...]", "Prompt text").action(async function (
-    this: Command,
-    promptParts: string[],
-  ) {
-    if (promptParts.length === 0 && process.stdin.isTTY) {
-      if (requestedJsonStrict) {
-        throw new InvalidArgumentError(
-          "Prompt is required (pass as argument, --file, or pipe via stdin)",
-        );
-      }
-      this.outputHelp();
-      return;
-    }
-
-    await handlePrompt(undefined, promptParts, {}, this, config);
+  configurePublicCli({
+    program,
+    argv: argv.slice(2),
+    config,
+    requestedJsonStrict,
+    topLevelVerbs: TOP_LEVEL_VERBS,
+    listBuiltInAgents,
+    detectAgentToken,
+    registerAgentCommand,
+    registerDefaultCommands,
+    handlePromptAction: async (command, promptParts) => {
+      await handlePrompt(undefined, promptParts, {}, command, config);
+    },
   });
-
-  program.addHelpText(
-    "after",
-    `
-Examples:
-  acpx codex sessions new
-  acpx codex "fix the tests"
-  acpx codex prompt "fix the tests"
-  acpx codex --no-wait "queue follow-up task"
-  acpx codex exec "what does this repo do"
-  acpx codex cancel
-  acpx codex set-mode plan
-  acpx codex set approval_policy conservative
-  acpx codex -s backend "fix the API"
-  acpx codex sessions
-  acpx codex sessions new --name backend
-  acpx codex sessions ensure --name backend
-  acpx codex sessions close backend
-  acpx codex status
-  acpx config show
-  acpx config init
-  acpx --ttl 30 codex "investigate flaky tests"
-  acpx claude "refactor auth"
-  acpx gemini "add logging"
-  acpx --agent ./my-custom-server "do something"`,
-  );
 
   program.exitOverride((error) => {
     throw error;
