+++ src/queue-ipc.ts
@@ -1,1517 +1,29 @@
-import type { SetSessionConfigOptionResponse } from "@agentclientprotocol/sdk";
-import { createHash, randomUUID } from "node:crypto";
-import fs from "node:fs/promises";
-import net from "node:net";
-import os from "node:os";
-import path from "node:path";
-import { normalizeOutputError } from "./error-normalization.js";
-import { QueueConnectionError, QueueProtocolError } from "./errors.js";
-import {
-  parseQueueOwnerMessage,
-  parseQueueRequest,
-  type QueueCancelRequest,
-  type QueueOwnerCancelResultMessage,
-  type QueueOwnerErrorMessage,
-  type QueueOwnerMessage,
-  type QueueOwnerSetConfigOptionResultMessage,
-  type QueueOwnerSetModeResultMessage,
-  type QueueRequest,
-  type QueueSetConfigOptionRequest,
-  type QueueSetModeRequest,
-  type QueueSubmitRequest,
-} from "./queue-messages.js";
-import type {
-  NonInteractivePermissionPolicy,
-  OutputErrorEmissionPolicy,
-  OutputFormatter,
-  PermissionMode,
-  SessionEnqueueResult,
-  SessionSendOutcome,
-} from "./types.js";
-
-const PROCESS_EXIT_GRACE_MS = 1_500;
-const PROCESS_POLL_MS = 50;
-const QUEUE_CONNECT_ATTEMPTS = 40;
-const QUEUE_OWNER_STALE_HEARTBEAT_MS = 15_000;
-export const QUEUE_CONNECT_RETRY_MS = 50;
-
-function queueBaseDir(): string {
-  return path.join(os.homedir(), ".acpx", "queues");
-}
-
-function makeQueueOwnerError(
-  requestId: string,
-  message: string,
-  detailCode: string,
-  options: {
-    retryable?: boolean;
-  } = {},
-): QueueOwnerErrorMessage {
-  return {
-    type: "error",
-    requestId,
-    code: "RUNTIME",
-    detailCode,
-    origin: "queue",
-    retryable: options.retryable,
-    message,
-  };
-}
-
-function makeQueueOwnerErrorFromUnknown(
-  requestId: string,
-  error: unknown,
-  detailCode: string,
-  options: {
-    retryable?: boolean;
-  } = {},
-): QueueOwnerErrorMessage {
-  const normalized = normalizeOutputError(error, {
-    defaultCode: "RUNTIME",
-    origin: "queue",
-    detailCode,
-    retryable: options.retryable,
-  });
-
-  return {
-    type: "error",
-    requestId,
-    code: normalized.code,
-    detailCode: normalized.detailCode,
-    origin: normalized.origin,
-    message: normalized.message,
-    retryable: normalized.retryable,
-    acp: normalized.acp,
-  };
-}
-
-export function isProcessAlive(pid: number | undefined): boolean {
-  if (!pid || !Number.isInteger(pid) || pid <= 0 || pid === process.pid) {
-    return false;
-  }
-
-  try {
-    process.kill(pid, 0);
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-async function waitForProcessExit(pid: number, timeoutMs: number): Promise<boolean> {
-  const deadline = Date.now() + Math.max(0, timeoutMs);
-  while (Date.now() <= deadline) {
-    if (!isProcessAlive(pid)) {
-      return true;
-    }
-    await new Promise<void>((resolve) => {
-      setTimeout(resolve, PROCESS_POLL_MS);
-    });
-  }
-
-  return !isProcessAlive(pid);
-}
-
-export async function terminateProcess(pid: number): Promise<boolean> {
-  if (!isProcessAlive(pid)) {
-    return false;
-  }
-
-  try {
-    process.kill(pid, "SIGTERM");
-  } catch {
-    return false;
-  }
-
-  if (await waitForProcessExit(pid, PROCESS_EXIT_GRACE_MS)) {
-    return true;
-  }
-
-  try {
-    process.kill(pid, "SIGKILL");
-  } catch {
-    return false;
-  }
-
-  await waitForProcessExit(pid, PROCESS_EXIT_GRACE_MS);
-  return true;
-}
-
-type QueueOwnerRecord = {
-  pid: number;
-  sessionId: string;
-  socketPath: string;
-  createdAt: string;
-  heartbeatAt: string;
-  ownerGeneration: number;
-  queueDepth: number;
-};
-
-export type QueueOwnerLease = {
-  sessionId: string;
-  lockPath: string;
-  socketPath: string;
-  createdAt: string;
-  ownerGeneration: number;
-};
-
+export {
+  ensureOwnerIsUsable,
+  isProcessAlive,
+  readQueueOwnerRecord,
+  readQueueOwnerStatus,
+  refreshQueueOwnerLease,
+  releaseQueueOwnerLease,
+  terminateProcess,
+  terminateQueueOwnerForSession,
+  tryAcquireQueueOwnerLease,
+  waitMs,
+  type QueueOwnerLease,
+  type QueueOwnerRecord,
+  type QueueOwnerStatus,
+} from "./queue-lease-store.js";
+export {
+  QUEUE_CONNECT_RETRY_MS,
+  tryCancelOnRunningOwner,
+  trySetConfigOptionOnRunningOwner,
+  trySetModeOnRunningOwner,
+  trySubmitToRunningOwner,
+  type SubmitToQueueOwnerOptions,
+} from "./queue-ipc-client.js";
+export {
+  SessionQueueOwner,
+  type QueueOwnerControlHandlers,
+  type QueueTask,
+} from "./queue-ipc-server.js";
 export type { QueueOwnerMessage, QueueSubmitRequest } from "./queue-messages.js";
-
-export type QueueTask = {
-  requestId: string;
-  message: string;
-  permissionMode: PermissionMode;
-  nonInteractivePermissions?: NonInteractivePermissionPolicy;
-  timeoutMs?: number;
-  suppressSdkConsoleErrors?: boolean;
-  waitForCompletion: boolean;
-  send: (message: QueueOwnerMessage) => void;
-  close: () => void;
-};
-
-function parseQueueOwnerRecord(raw: unknown): QueueOwnerRecord | null {
-  if (!raw || typeof raw !== "object" || Array.isArray(raw)) {
-    return null;
-  }
-  const record = raw as Record<string, unknown>;
-
-  if (
-    !Number.isInteger(record.pid) ||
-    (record.pid as number) <= 0 ||
-    typeof record.sessionId !== "string" ||
-    typeof record.socketPath !== "string" ||
-    typeof record.createdAt !== "string" ||
-    typeof record.heartbeatAt !== "string" ||
-    !Number.isInteger(record.ownerGeneration) ||
-    (record.ownerGeneration as number) <= 0 ||
-    !Number.isInteger(record.queueDepth) ||
-    (record.queueDepth as number) < 0
-  ) {
-    return null;
-  }
-
-  return {
-    pid: record.pid as number,
-    sessionId: record.sessionId,
-    socketPath: record.socketPath,
-    createdAt: record.createdAt,
-    heartbeatAt: record.heartbeatAt,
-    ownerGeneration: record.ownerGeneration as number,
-    queueDepth: record.queueDepth as number,
-  };
-}
-
-function createOwnerGeneration(): number {
-  return Date.now() * 1_000 + Math.floor(Math.random() * 1_000);
-}
-
-function nowIso(): string {
-  return new Date().toISOString();
-}
-
-function isQueueOwnerHeartbeatStale(owner: QueueOwnerRecord): boolean {
-  const heartbeatMs = Date.parse(owner.heartbeatAt);
-  if (!Number.isFinite(heartbeatMs)) {
-    return true;
-  }
-  return Date.now() - heartbeatMs > QUEUE_OWNER_STALE_HEARTBEAT_MS;
-}
-
-function queueKeyForSession(sessionId: string): string {
-  return createHash("sha256").update(sessionId).digest("hex").slice(0, 24);
-}
-
-function queueLockFilePath(sessionId: string): string {
-  return path.join(queueBaseDir(), `${queueKeyForSession(sessionId)}.lock`);
-}
-
-function queueSocketPath(sessionId: string): string {
-  const key = queueKeyForSession(sessionId);
-  if (process.platform === "win32") {
-    return `\\\\.\\pipe\\acpx-${key}`;
-  }
-  return path.join(queueBaseDir(), `${key}.sock`);
-}
-
-async function ensureQueueDir(): Promise<void> {
-  await fs.mkdir(queueBaseDir(), { recursive: true });
-}
-
-async function removeSocketFile(socketPath: string): Promise<void> {
-  if (process.platform === "win32") {
-    return;
-  }
-
-  try {
-    await fs.unlink(socketPath);
-  } catch (error) {
-    if ((error as NodeJS.ErrnoException).code !== "ENOENT") {
-      throw error;
-    }
-  }
-}
-
-async function readQueueOwnerRecord(
-  sessionId: string,
-): Promise<QueueOwnerRecord | undefined> {
-  const lockPath = queueLockFilePath(sessionId);
-  try {
-    const payload = await fs.readFile(lockPath, "utf8");
-    const parsed = parseQueueOwnerRecord(JSON.parse(payload));
-    return parsed ?? undefined;
-  } catch {
-    return undefined;
-  }
-}
-
-export async function readQueueOwnerStatus(sessionId: string): Promise<
-  | {
-      pid: number;
-      socketPath: string;
-      heartbeatAt: string;
-      ownerGeneration: number;
-      queueDepth: number;
-      alive: boolean;
-      stale: boolean;
-    }
-  | undefined
-> {
-  const owner = await readQueueOwnerRecord(sessionId);
-  if (!owner) {
-    return undefined;
-  }
-
-  const alive = await ensureOwnerIsUsable(sessionId, owner);
-  if (!alive) {
-    return undefined;
-  }
-
-  return {
-    pid: owner.pid,
-    socketPath: owner.socketPath,
-    heartbeatAt: owner.heartbeatAt,
-    ownerGeneration: owner.ownerGeneration,
-    queueDepth: owner.queueDepth,
-    alive,
-    stale: isQueueOwnerHeartbeatStale(owner),
-  };
-}
-
-async function ensureOwnerIsUsable(
-  sessionId: string,
-  owner: QueueOwnerRecord,
-): Promise<boolean> {
-  const alive = isProcessAlive(owner.pid);
-  const stale = isQueueOwnerHeartbeatStale(owner);
-  if (alive && !stale) {
-    return true;
-  }
-
-  if (alive) {
-    await terminateProcess(owner.pid).catch(() => {
-      // best effort stale owner termination
-    });
-  }
-  await cleanupStaleQueueOwner(sessionId, owner);
-  return false;
-}
-
-async function cleanupStaleQueueOwner(
-  sessionId: string,
-  owner: QueueOwnerRecord | undefined,
-): Promise<void> {
-  const lockPath = queueLockFilePath(sessionId);
-  const socketPath = owner?.socketPath ?? queueSocketPath(sessionId);
-
-  await removeSocketFile(socketPath).catch(() => {
-    // ignore stale socket cleanup failures
-  });
-
-  await fs.unlink(lockPath).catch((error) => {
-    if ((error as NodeJS.ErrnoException).code !== "ENOENT") {
-      throw error;
-    }
-  });
-}
-
-export async function tryAcquireQueueOwnerLease(
-  sessionId: string,
-  nowIsoFactory: () => string = nowIso,
-): Promise<QueueOwnerLease | undefined> {
-  await ensureQueueDir();
-  const lockPath = queueLockFilePath(sessionId);
-  const socketPath = queueSocketPath(sessionId);
-  const createdAt = nowIsoFactory();
-  const ownerGeneration = createOwnerGeneration();
-  const payload = JSON.stringify(
-    {
-      pid: process.pid,
-      sessionId,
-      socketPath,
-      createdAt,
-      heartbeatAt: createdAt,
-      ownerGeneration,
-      queueDepth: 0,
-    },
-    null,
-    2,
-  );
-
-  try {
-    await fs.writeFile(lockPath, `${payload}\n`, {
-      encoding: "utf8",
-      flag: "wx",
-    });
-    await removeSocketFile(socketPath).catch(() => {
-      // best-effort stale socket cleanup after ownership is acquired
-    });
-    return {
-      sessionId,
-      lockPath,
-      socketPath,
-      createdAt,
-      ownerGeneration,
-    };
-  } catch (error) {
-    if ((error as NodeJS.ErrnoException).code !== "EEXIST") {
-      throw error;
-    }
-
-    const owner = await readQueueOwnerRecord(sessionId);
-    if (!owner) {
-      await cleanupStaleQueueOwner(sessionId, owner);
-      return undefined;
-    }
-
-    if (!isProcessAlive(owner.pid) || isQueueOwnerHeartbeatStale(owner)) {
-      if (isProcessAlive(owner.pid)) {
-        await terminateProcess(owner.pid).catch(() => {
-          // best effort stale owner termination
-        });
-      }
-      await cleanupStaleQueueOwner(sessionId, owner);
-    }
-    return undefined;
-  }
-}
-
-export async function refreshQueueOwnerLease(
-  lease: QueueOwnerLease,
-  options: {
-    queueDepth: number;
-  },
-  nowIsoFactory: () => string = nowIso,
-): Promise<void> {
-  const payload = JSON.stringify(
-    {
-      pid: process.pid,
-      sessionId: lease.sessionId,
-      socketPath: lease.socketPath,
-      createdAt: lease.createdAt,
-      heartbeatAt: nowIsoFactory(),
-      ownerGeneration: lease.ownerGeneration,
-      queueDepth: Math.max(0, Math.round(options.queueDepth)),
-    },
-    null,
-    2,
-  );
-  await fs.writeFile(lease.lockPath, `${payload}\n`, {
-    encoding: "utf8",
-  });
-}
-
-export async function releaseQueueOwnerLease(lease: QueueOwnerLease): Promise<void> {
-  await removeSocketFile(lease.socketPath).catch(() => {
-    // ignore best-effort cleanup failures
-  });
-
-  await fs.unlink(lease.lockPath).catch((error) => {
-    if ((error as NodeJS.ErrnoException).code !== "ENOENT") {
-      throw error;
-    }
-  });
-}
-
-function shouldRetryQueueConnect(error: unknown): boolean {
-  const code = (error as NodeJS.ErrnoException).code;
-  return code === "ENOENT" || code === "ECONNREFUSED";
-}
-
-export async function waitMs(ms: number): Promise<void> {
-  await new Promise<void>((resolve) => {
-    setTimeout(resolve, ms);
-  });
-}
-
-async function connectToSocket(socketPath: string): Promise<net.Socket> {
-  return await new Promise<net.Socket>((resolve, reject) => {
-    const socket = net.createConnection(socketPath);
-
-    const onConnect = () => {
-      socket.off("error", onError);
-      resolve(socket);
-    };
-    const onError = (error: Error) => {
-      socket.off("connect", onConnect);
-      reject(error);
-    };
-
-    socket.once("connect", onConnect);
-    socket.once("error", onError);
-  });
-}
-
-async function connectToQueueOwner(
-  owner: QueueOwnerRecord,
-): Promise<net.Socket | undefined> {
-  let lastError: unknown;
-
-  for (let attempt = 0; attempt < QUEUE_CONNECT_ATTEMPTS; attempt += 1) {
-    try {
-      return await connectToSocket(owner.socketPath);
-    } catch (error) {
-      lastError = error;
-      if (!shouldRetryQueueConnect(error)) {
-        throw error;
-      }
-
-      if (!isProcessAlive(owner.pid)) {
-        return undefined;
-      }
-      await waitMs(QUEUE_CONNECT_RETRY_MS);
-    }
-  }
-
-  if (lastError && !shouldRetryQueueConnect(lastError)) {
-    throw lastError;
-  }
-
-  return undefined;
-}
-
-function writeQueueMessage(socket: net.Socket, message: QueueOwnerMessage): void {
-  if (socket.destroyed || !socket.writable) {
-    return;
-  }
-  socket.write(`${JSON.stringify(message)}\n`);
-}
-
-export type QueueOwnerControlHandlers = {
-  cancelPrompt: () => Promise<boolean>;
-  setSessionMode: (modeId: string, timeoutMs?: number) => Promise<void>;
-  setSessionConfigOption: (
-    configId: string,
-    value: string,
-    timeoutMs?: number,
-  ) => Promise<SetSessionConfigOptionResponse>;
-};
-
-export class SessionQueueOwner {
-  private readonly server: net.Server;
-  private readonly controlHandlers: QueueOwnerControlHandlers;
-  private readonly pending: QueueTask[] = [];
-  private readonly waiters: Array<(task: QueueTask | undefined) => void> = [];
-  private closed = false;
-
-  private constructor(server: net.Server, controlHandlers: QueueOwnerControlHandlers) {
-    this.server = server;
-    this.controlHandlers = controlHandlers;
-  }
-
-  static async start(
-    lease: QueueOwnerLease,
-    controlHandlers: QueueOwnerControlHandlers,
-  ): Promise<SessionQueueOwner> {
-    const ownerRef: { current: SessionQueueOwner | undefined } = { current: undefined };
-    const server = net.createServer((socket) => {
-      ownerRef.current?.handleConnection(socket);
-    });
-    ownerRef.current = new SessionQueueOwner(server, controlHandlers);
-
-    await new Promise<void>((resolve, reject) => {
-      const onListening = () => {
-        server.off("error", onError);
-        resolve();
-      };
-      const onError = (error: Error) => {
-        server.off("listening", onListening);
-        reject(error);
-      };
-
-      server.once("listening", onListening);
-      server.once("error", onError);
-      server.listen(lease.socketPath);
-    });
-
-    return ownerRef.current!;
-  }
-
-  async close(): Promise<void> {
-    if (this.closed) {
-      return;
-    }
-
-    this.closed = true;
-    for (const waiter of this.waiters.splice(0)) {
-      waiter(undefined);
-    }
-
-    for (const task of this.pending.splice(0)) {
-      if (task.waitForCompletion) {
-        task.send(
-          makeQueueOwnerError(
-            task.requestId,
-            "Queue owner shutting down before prompt execution",
-            "QUEUE_OWNER_SHUTTING_DOWN",
-            {
-              retryable: true,
-            },
-          ),
-        );
-      }
-      task.close();
-    }
-
-    await new Promise<void>((resolve) => {
-      this.server.close(() => resolve());
-    });
-  }
-
-  async nextTask(timeoutMs?: number): Promise<QueueTask | undefined> {
-    if (this.pending.length > 0) {
-      return this.pending.shift();
-    }
-    if (this.closed) {
-      return undefined;
-    }
-
-    return await new Promise<QueueTask | undefined>((resolve) => {
-      const shouldTimeout = timeoutMs != null;
-      const timer =
-        shouldTimeout &&
-        setTimeout(
-          () => {
-            const index = this.waiters.indexOf(waiter);
-            if (index >= 0) {
-              this.waiters.splice(index, 1);
-            }
-            resolve(undefined);
-          },
-          Math.max(0, timeoutMs),
-        );
-
-      const waiter = (task: QueueTask | undefined) => {
-        if (timer) {
-          clearTimeout(timer);
-        }
-        resolve(task);
-      };
-
-      this.waiters.push(waiter);
-    });
-  }
-
-  queueDepth(): number {
-    return this.pending.length;
-  }
-
-  private enqueue(task: QueueTask): void {
-    if (this.closed) {
-      if (task.waitForCompletion) {
-        task.send(
-          makeQueueOwnerError(
-            task.requestId,
-            "Queue owner is shutting down",
-            "QUEUE_OWNER_SHUTTING_DOWN",
-            {
-              retryable: true,
-            },
-          ),
-        );
-      }
-      task.close();
-      return;
-    }
-
-    const waiter = this.waiters.shift();
-    if (waiter) {
-      waiter(task);
-      return;
-    }
-
-    this.pending.push(task);
-  }
-
-  private handleConnection(socket: net.Socket): void {
-    socket.setEncoding("utf8");
-
-    if (this.closed) {
-      writeQueueMessage(
-        socket,
-        makeQueueOwnerError("unknown", "Queue owner is closed", "QUEUE_OWNER_CLOSED", {
-          retryable: true,
-        }),
-      );
-      socket.end();
-      return;
-    }
-
-    let buffer = "";
-    let handled = false;
-
-    const fail = (requestId: string, message: string, detailCode: string): void => {
-      writeQueueMessage(
-        socket,
-        makeQueueOwnerError(requestId, message, detailCode, {
-          retryable: false,
-        }),
-      );
-      socket.end();
-    };
-
-    const processLine = (line: string): void => {
-      if (handled) {
-        return;
-      }
-      handled = true;
-
-      let parsed: unknown;
-      try {
-        parsed = JSON.parse(line);
-      } catch {
-        fail(
-          "unknown",
-          "Invalid queue request payload",
-          "QUEUE_REQUEST_PAYLOAD_INVALID_JSON",
-        );
-        return;
-      }
-
-      const request = parseQueueRequest(parsed);
-      if (!request) {
-        fail("unknown", "Invalid queue request", "QUEUE_REQUEST_INVALID");
-        return;
-      }
-
-      if (request.type === "cancel_prompt") {
-        writeQueueMessage(socket, {
-          type: "accepted",
-          requestId: request.requestId,
-        });
-        void this.controlHandlers
-          .cancelPrompt()
-          .then((cancelled) => {
-            writeQueueMessage(socket, {
-              type: "cancel_result",
-              requestId: request.requestId,
-              cancelled,
-            });
-          })
-          .catch((error) => {
-            writeQueueMessage(
-              socket,
-              makeQueueOwnerErrorFromUnknown(
-                request.requestId,
-                error,
-                "QUEUE_CONTROL_REQUEST_FAILED",
-              ),
-            );
-          })
-          .finally(() => {
-            if (!socket.destroyed) {
-              socket.end();
-            }
-          });
-        return;
-      }
-
-      if (request.type === "set_mode") {
-        writeQueueMessage(socket, {
-          type: "accepted",
-          requestId: request.requestId,
-        });
-        void this.controlHandlers
-          .setSessionMode(request.modeId, request.timeoutMs)
-          .then(() => {
-            writeQueueMessage(socket, {
-              type: "set_mode_result",
-              requestId: request.requestId,
-              modeId: request.modeId,
-            });
-          })
-          .catch((error) => {
-            writeQueueMessage(
-              socket,
-              makeQueueOwnerErrorFromUnknown(
-                request.requestId,
-                error,
-                "QUEUE_CONTROL_REQUEST_FAILED",
-              ),
-            );
-          })
-          .finally(() => {
-            if (!socket.destroyed) {
-              socket.end();
-            }
-          });
-        return;
-      }
-
-      if (request.type === "set_config_option") {
-        writeQueueMessage(socket, {
-          type: "accepted",
-          requestId: request.requestId,
-        });
-        void this.controlHandlers
-          .setSessionConfigOption(request.configId, request.value, request.timeoutMs)
-          .then((response) => {
-            writeQueueMessage(socket, {
-              type: "set_config_option_result",
-              requestId: request.requestId,
-              response,
-            });
-          })
-          .catch((error) => {
-            writeQueueMessage(
-              socket,
-              makeQueueOwnerErrorFromUnknown(
-                request.requestId,
-                error,
-                "QUEUE_CONTROL_REQUEST_FAILED",
-              ),
-            );
-          })
-          .finally(() => {
-            if (!socket.destroyed) {
-              socket.end();
-            }
-          });
-        return;
-      }
-
-      const task: QueueTask = {
-        requestId: request.requestId,
-        message: request.message,
-        permissionMode: request.permissionMode,
-        nonInteractivePermissions: request.nonInteractivePermissions,
-        timeoutMs: request.timeoutMs,
-        suppressSdkConsoleErrors: request.suppressSdkConsoleErrors,
-        waitForCompletion: request.waitForCompletion,
-        send: (message) => {
-          writeQueueMessage(socket, message);
-        },
-        close: () => {
-          if (!socket.destroyed) {
-            socket.end();
-          }
-        },
-      };
-
-      writeQueueMessage(socket, {
-        type: "accepted",
-        requestId: request.requestId,
-      });
-
-      if (!request.waitForCompletion) {
-        task.close();
-      }
-
-      this.enqueue(task);
-    };
-
-    socket.on("data", (chunk: string) => {
-      buffer += chunk;
-
-      let index = buffer.indexOf("\n");
-      while (index >= 0) {
-        const line = buffer.slice(0, index).trim();
-        buffer = buffer.slice(index + 1);
-
-        if (line.length > 0) {
-          processLine(line);
-        }
-
-        index = buffer.indexOf("\n");
-      }
-    });
-
-    socket.on("error", () => {
-      // no-op: queue processing continues even if client disconnects
-    });
-  }
-}
-
-export type SubmitToQueueOwnerOptions = {
-  sessionId: string;
-  message: string;
-  permissionMode: PermissionMode;
-  nonInteractivePermissions?: NonInteractivePermissionPolicy;
-  outputFormatter: OutputFormatter;
-  errorEmissionPolicy?: OutputErrorEmissionPolicy;
-  timeoutMs?: number;
-  suppressSdkConsoleErrors?: boolean;
-  waitForCompletion: boolean;
-  verbose?: boolean;
-};
-
-async function submitToQueueOwner(
-  owner: QueueOwnerRecord,
-  options: SubmitToQueueOwnerOptions,
-): Promise<SessionSendOutcome | undefined> {
-  const socket = await connectToQueueOwner(owner);
-  if (!socket) {
-    return undefined;
-  }
-
-  socket.setEncoding("utf8");
-  const requestId = randomUUID();
-  const request: QueueSubmitRequest = {
-    type: "submit_prompt",
-    requestId,
-    message: options.message,
-    permissionMode: options.permissionMode,
-    nonInteractivePermissions: options.nonInteractivePermissions,
-    timeoutMs: options.timeoutMs,
-    suppressSdkConsoleErrors: options.suppressSdkConsoleErrors,
-    waitForCompletion: options.waitForCompletion,
-  };
-
-  options.outputFormatter.setContext({
-    sessionId: options.sessionId,
-    requestId,
-    stream: "prompt",
-  });
-
-  return await new Promise<SessionSendOutcome>((resolve, reject) => {
-    let settled = false;
-    let acknowledged = false;
-    let buffer = "";
-    let sawDone = false;
-
-    const finishResolve = (result: SessionSendOutcome) => {
-      if (settled) {
-        return;
-      }
-      settled = true;
-      socket.removeAllListeners();
-      if (!socket.destroyed) {
-        socket.end();
-      }
-      resolve(result);
-    };
-
-    const finishReject = (error: unknown) => {
-      if (settled) {
-        return;
-      }
-      settled = true;
-      socket.removeAllListeners();
-      if (!socket.destroyed) {
-        socket.destroy();
-      }
-      reject(error);
-    };
-
-    const processLine = (line: string): void => {
-      let parsed: unknown;
-      try {
-        parsed = JSON.parse(line);
-      } catch {
-        finishReject(
-          new QueueProtocolError("Queue owner sent invalid JSON payload", {
-            detailCode: "QUEUE_PROTOCOL_INVALID_JSON",
-            origin: "queue",
-            retryable: true,
-          }),
-        );
-        return;
-      }
-
-      const message = parseQueueOwnerMessage(parsed);
-      if (!message || message.requestId !== requestId) {
-        finishReject(
-          new QueueProtocolError("Queue owner sent malformed message", {
-            detailCode: "QUEUE_PROTOCOL_MALFORMED_MESSAGE",
-            origin: "queue",
-            retryable: true,
-          }),
-        );
-        return;
-      }
-
-      if (message.type === "accepted") {
-        acknowledged = true;
-        options.outputFormatter.setContext({
-          sessionId: options.sessionId,
-          requestId: message.requestId,
-          stream: "prompt",
-        });
-        if (!options.waitForCompletion) {
-          const queued: SessionEnqueueResult = {
-            queued: true,
-            sessionId: options.sessionId,
-            requestId,
-          };
-          finishResolve(queued);
-        }
-        return;
-      }
-
-      if (message.type === "error") {
-        options.outputFormatter.setContext({
-          sessionId: options.sessionId,
-          requestId: message.requestId,
-          stream: "prompt",
-        });
-        options.outputFormatter.onError({
-          code: message.code ?? "RUNTIME",
-          detailCode: message.detailCode,
-          origin: message.origin ?? "queue",
-          message: message.message,
-          retryable: message.retryable,
-          acp: message.acp,
-        });
-        options.outputFormatter.flush();
-        const queueErrorAlreadyEmitted =
-          options.errorEmissionPolicy?.queueErrorAlreadyEmitted ?? true;
-        finishReject(
-          new QueueConnectionError(message.message, {
-            outputCode: message.code,
-            detailCode: message.detailCode,
-            origin: message.origin ?? "queue",
-            retryable: message.retryable,
-            acp: message.acp,
-            ...(queueErrorAlreadyEmitted ? { outputAlreadyEmitted: true } : {}),
-          }),
-        );
-        return;
-      }
-
-      if (!acknowledged) {
-        finishReject(
-          new QueueConnectionError("Queue owner did not acknowledge request", {
-            detailCode: "QUEUE_ACK_MISSING",
-            origin: "queue",
-            retryable: true,
-          }),
-        );
-        return;
-      }
-
-      if (message.type === "session_update") {
-        options.outputFormatter.onSessionUpdate(message.notification);
-        return;
-      }
-
-      if (message.type === "client_operation") {
-        options.outputFormatter.onClientOperation(message.operation);
-        return;
-      }
-
-      if (message.type === "done") {
-        options.outputFormatter.onDone(message.stopReason);
-        sawDone = true;
-        return;
-      }
-
-      if (message.type === "result") {
-        if (!sawDone) {
-          options.outputFormatter.onDone(message.result.stopReason);
-        }
-        options.outputFormatter.flush();
-        finishResolve(message.result);
-        return;
-      }
-
-      finishReject(
-        new QueueProtocolError("Queue owner returned unexpected response", {
-          detailCode: "QUEUE_PROTOCOL_UNEXPECTED_RESPONSE",
-          origin: "queue",
-          retryable: true,
-        }),
-      );
-    };
-
-    socket.on("data", (chunk: string) => {
-      buffer += chunk;
-
-      let index = buffer.indexOf("\n");
-      while (index >= 0) {
-        const line = buffer.slice(0, index).trim();
-        buffer = buffer.slice(index + 1);
-
-        if (line.length > 0) {
-          processLine(line);
-        }
-
-        index = buffer.indexOf("\n");
-      }
-    });
-
-    socket.once("error", (error) => {
-      finishReject(error);
-    });
-
-    socket.once("close", () => {
-      if (settled) {
-        return;
-      }
-
-      if (!acknowledged) {
-        finishReject(
-          new QueueConnectionError(
-            "Queue owner disconnected before acknowledging request",
-            {
-              detailCode: "QUEUE_DISCONNECTED_BEFORE_ACK",
-              origin: "queue",
-              retryable: true,
-            },
-          ),
-        );
-        return;
-      }
-
-      if (!options.waitForCompletion) {
-        const queued: SessionEnqueueResult = {
-          queued: true,
-          sessionId: options.sessionId,
-          requestId,
-        };
-        finishResolve(queued);
-        return;
-      }
-
-      finishReject(
-        new QueueConnectionError("Queue owner disconnected before prompt completion", {
-          detailCode: "QUEUE_DISCONNECTED_BEFORE_COMPLETION",
-          origin: "queue",
-          retryable: true,
-        }),
-      );
-    });
-
-    socket.write(`${JSON.stringify(request)}\n`);
-  });
-}
-
-async function submitControlToQueueOwner<TResponse extends QueueOwnerMessage>(
-  owner: QueueOwnerRecord,
-  request: QueueRequest,
-  isExpectedResponse: (message: QueueOwnerMessage) => message is TResponse,
-): Promise<TResponse | undefined> {
-  const socket = await connectToQueueOwner(owner);
-  if (!socket) {
-    return undefined;
-  }
-
-  socket.setEncoding("utf8");
-
-  return await new Promise<TResponse>((resolve, reject) => {
-    let settled = false;
-    let acknowledged = false;
-    let buffer = "";
-
-    const finishResolve = (result: TResponse) => {
-      if (settled) {
-        return;
-      }
-      settled = true;
-      socket.removeAllListeners();
-      if (!socket.destroyed) {
-        socket.end();
-      }
-      resolve(result);
-    };
-
-    const finishReject = (error: unknown) => {
-      if (settled) {
-        return;
-      }
-      settled = true;
-      socket.removeAllListeners();
-      if (!socket.destroyed) {
-        socket.destroy();
-      }
-      reject(error);
-    };
-
-    const processLine = (line: string): void => {
-      let parsed: unknown;
-      try {
-        parsed = JSON.parse(line);
-      } catch {
-        finishReject(
-          new QueueProtocolError("Queue owner sent invalid JSON payload", {
-            detailCode: "QUEUE_PROTOCOL_INVALID_JSON",
-            origin: "queue",
-            retryable: true,
-          }),
-        );
-        return;
-      }
-
-      const message = parseQueueOwnerMessage(parsed);
-      if (!message || message.requestId !== request.requestId) {
-        finishReject(
-          new QueueProtocolError("Queue owner sent malformed message", {
-            detailCode: "QUEUE_PROTOCOL_MALFORMED_MESSAGE",
-            origin: "queue",
-            retryable: true,
-          }),
-        );
-        return;
-      }
-
-      if (message.type === "accepted") {
-        acknowledged = true;
-        return;
-      }
-
-      if (message.type === "error") {
-        finishReject(
-          new QueueConnectionError(message.message, {
-            outputCode: message.code,
-            detailCode: message.detailCode,
-            origin: message.origin ?? "queue",
-            retryable: message.retryable,
-            acp: message.acp,
-          }),
-        );
-        return;
-      }
-
-      if (!acknowledged) {
-        finishReject(
-          new QueueConnectionError("Queue owner did not acknowledge request", {
-            detailCode: "QUEUE_ACK_MISSING",
-            origin: "queue",
-            retryable: true,
-          }),
-        );
-        return;
-      }
-
-      if (!isExpectedResponse(message)) {
-        finishReject(
-          new QueueProtocolError("Queue owner returned unexpected response", {
-            detailCode: "QUEUE_PROTOCOL_UNEXPECTED_RESPONSE",
-            origin: "queue",
-            retryable: true,
-          }),
-        );
-        return;
-      }
-
-      finishResolve(message);
-    };
-
-    socket.on("data", (chunk: string) => {
-      buffer += chunk;
-
-      let index = buffer.indexOf("\n");
-      while (index >= 0) {
-        const line = buffer.slice(0, index).trim();
-        buffer = buffer.slice(index + 1);
-
-        if (line.length > 0) {
-          processLine(line);
-        }
-
-        index = buffer.indexOf("\n");
-      }
-    });
-
-    socket.once("error", (error) => {
-      finishReject(error);
-    });
-
-    socket.once("close", () => {
-      if (settled) {
-        return;
-      }
-      if (!acknowledged) {
-        finishReject(
-          new QueueConnectionError(
-            "Queue owner disconnected before acknowledging request",
-            {
-              detailCode: "QUEUE_DISCONNECTED_BEFORE_ACK",
-              origin: "queue",
-              retryable: true,
-            },
-          ),
-        );
-        return;
-      }
-      finishReject(
-        new QueueConnectionError("Queue owner disconnected before responding", {
-          detailCode: "QUEUE_DISCONNECTED_BEFORE_COMPLETION",
-          origin: "queue",
-          retryable: true,
-        }),
-      );
-    });
-
-    socket.write(`${JSON.stringify(request)}\n`);
-  });
-}
-
-async function submitCancelToQueueOwner(
-  owner: QueueOwnerRecord,
-): Promise<boolean | undefined> {
-  const request: QueueCancelRequest = {
-    type: "cancel_prompt",
-    requestId: randomUUID(),
-  };
-  const response = await submitControlToQueueOwner(
-    owner,
-    request,
-    (message): message is QueueOwnerCancelResultMessage =>
-      message.type === "cancel_result",
-  );
-  if (!response) {
-    return undefined;
-  }
-  if (response.requestId !== request.requestId) {
-    throw new QueueProtocolError("Queue owner returned mismatched cancel response", {
-      detailCode: "QUEUE_PROTOCOL_MALFORMED_MESSAGE",
-      origin: "queue",
-      retryable: true,
-    });
-  }
-  return response.cancelled;
-}
-
-async function submitSetModeToQueueOwner(
-  owner: QueueOwnerRecord,
-  modeId: string,
-  timeoutMs?: number,
-): Promise<boolean | undefined> {
-  const request: QueueSetModeRequest = {
-    type: "set_mode",
-    requestId: randomUUID(),
-    modeId,
-    timeoutMs,
-  };
-  const response = await submitControlToQueueOwner(
-    owner,
-    request,
-    (message): message is QueueOwnerSetModeResultMessage =>
-      message.type === "set_mode_result",
-  );
-  if (!response) {
-    return undefined;
-  }
-  if (response.requestId !== request.requestId) {
-    throw new QueueProtocolError("Queue owner returned mismatched set_mode response", {
-      detailCode: "QUEUE_PROTOCOL_MALFORMED_MESSAGE",
-      origin: "queue",
-      retryable: true,
-    });
-  }
-  return true;
-}
-
-async function submitSetConfigOptionToQueueOwner(
-  owner: QueueOwnerRecord,
-  configId: string,
-  value: string,
-  timeoutMs?: number,
-): Promise<SetSessionConfigOptionResponse | undefined> {
-  const request: QueueSetConfigOptionRequest = {
-    type: "set_config_option",
-    requestId: randomUUID(),
-    configId,
-    value,
-    timeoutMs,
-  };
-  const response = await submitControlToQueueOwner(
-    owner,
-    request,
-    (message): message is QueueOwnerSetConfigOptionResultMessage =>
-      message.type === "set_config_option_result",
-  );
-  if (!response) {
-    return undefined;
-  }
-  if (response.requestId !== request.requestId) {
-    throw new QueueProtocolError(
-      "Queue owner returned mismatched set_config_option response",
-      {
-        detailCode: "QUEUE_PROTOCOL_MALFORMED_MESSAGE",
-        origin: "queue",
-        retryable: true,
-      },
-    );
-  }
-  return response.response;
-}
-
-export async function trySubmitToRunningOwner(
-  options: SubmitToQueueOwnerOptions,
-): Promise<SessionSendOutcome | undefined> {
-  const owner = await readQueueOwnerRecord(options.sessionId);
-  if (!owner) {
-    return undefined;
-  }
-
-  if (!(await ensureOwnerIsUsable(options.sessionId, owner))) {
-    return undefined;
-  }
-
-  const submitted = await submitToQueueOwner(owner, options);
-  if (submitted) {
-    if (options.verbose) {
-      process.stderr.write(
-        `[acpx] queued prompt on active owner pid ${owner.pid} for session ${options.sessionId}\n`,
-      );
-    }
-    return submitted;
-  }
-
-  if (!(await ensureOwnerIsUsable(options.sessionId, owner))) {
-    return undefined;
-  }
-
-  throw new QueueConnectionError(
-    "Session queue owner is running but not accepting queue requests",
-    {
-      detailCode: "QUEUE_NOT_ACCEPTING_REQUESTS",
-      origin: "queue",
-      retryable: true,
-    },
-  );
-}
-
-export async function tryCancelOnRunningOwner(options: {
-  sessionId: string;
-  verbose?: boolean;
-}): Promise<boolean | undefined> {
-  const owner = await readQueueOwnerRecord(options.sessionId);
-  if (!owner) {
-    return undefined;
-  }
-
-  if (!(await ensureOwnerIsUsable(options.sessionId, owner))) {
-    return undefined;
-  }
-
-  const cancelled = await submitCancelToQueueOwner(owner);
-  if (cancelled !== undefined) {
-    if (options.verbose) {
-      process.stderr.write(
-        `[acpx] requested cancel on active owner pid ${owner.pid} for session ${options.sessionId}\n`,
-      );
-    }
-    return cancelled;
-  }
-
-  if (!(await ensureOwnerIsUsable(options.sessionId, owner))) {
-    return undefined;
-  }
-
-  throw new QueueConnectionError(
-    "Session queue owner is running but not accepting cancel requests",
-    {
-      detailCode: "QUEUE_NOT_ACCEPTING_REQUESTS",
-      origin: "queue",
-      retryable: true,
-    },
-  );
-}
-
-export async function trySetModeOnRunningOwner(
-  sessionId: string,
-  modeId: string,
-  timeoutMs: number | undefined,
-  verbose: boolean | undefined,
-): Promise<boolean | undefined> {
-  const owner = await readQueueOwnerRecord(sessionId);
-  if (!owner) {
-    return undefined;
-  }
-
-  if (!(await ensureOwnerIsUsable(sessionId, owner))) {
-    return undefined;
-  }
-
-  const submitted = await submitSetModeToQueueOwner(owner, modeId, timeoutMs);
-  if (submitted) {
-    if (verbose) {
-      process.stderr.write(
-        `[acpx] requested session/set_mode on owner pid ${owner.pid} for session ${sessionId}\n`,
-      );
-    }
-    return true;
-  }
-
-  if (!(await ensureOwnerIsUsable(sessionId, owner))) {
-    return undefined;
-  }
-
-  throw new QueueConnectionError(
-    "Session queue owner is running but not accepting set_mode requests",
-    {
-      detailCode: "QUEUE_NOT_ACCEPTING_REQUESTS",
-      origin: "queue",
-      retryable: true,
-    },
-  );
-}
-
-export async function trySetConfigOptionOnRunningOwner(
-  sessionId: string,
-  configId: string,
-  value: string,
-  timeoutMs: number | undefined,
-  verbose: boolean | undefined,
-): Promise<SetSessionConfigOptionResponse | undefined> {
-  const owner = await readQueueOwnerRecord(sessionId);
-  if (!owner) {
-    return undefined;
-  }
-
-  if (!(await ensureOwnerIsUsable(sessionId, owner))) {
-    return undefined;
-  }
-
-  const response = await submitSetConfigOptionToQueueOwner(
-    owner,
-    configId,
-    value,
-    timeoutMs,
-  );
-  if (response) {
-    if (verbose) {
-      process.stderr.write(
-        `[acpx] requested session/set_config_option on owner pid ${owner.pid} for session ${sessionId}\n`,
-      );
-    }
-    return response;
-  }
-
-  if (!(await ensureOwnerIsUsable(sessionId, owner))) {
-    return undefined;
-  }
-
-  throw new QueueConnectionError(
-    "Session queue owner is running but not accepting set_config_option requests",
-    {
-      detailCode: "QUEUE_NOT_ACCEPTING_REQUESTS",
-      origin: "queue",
-      retryable: true,
-    },
-  );
-}
-
-export async function terminateQueueOwnerForSession(sessionId: string): Promise<void> {
-  const owner = await readQueueOwnerRecord(sessionId);
-  if (!owner) {
-    return;
-  }
-
-  if (isProcessAlive(owner.pid)) {
-    await terminateProcess(owner.pid);
-  }
-
-  await cleanupStaleQueueOwner(sessionId, owner);
-}
