+++ src/queue-lease-store.ts
@@ -1,0 +1,363 @@
+import { createHash } from "node:crypto";
+import fs from "node:fs/promises";
+import os from "node:os";
+import path from "node:path";
+
+const PROCESS_EXIT_GRACE_MS = 1_500;
+const PROCESS_POLL_MS = 50;
+const QUEUE_OWNER_STALE_HEARTBEAT_MS = 15_000;
+
+export type QueueOwnerRecord = {
+  pid: number;
+  sessionId: string;
+  socketPath: string;
+  createdAt: string;
+  heartbeatAt: string;
+  ownerGeneration: number;
+  queueDepth: number;
+};
+
+export type QueueOwnerLease = {
+  sessionId: string;
+  lockPath: string;
+  socketPath: string;
+  createdAt: string;
+  ownerGeneration: number;
+};
+
+export type QueueOwnerStatus = {
+  pid: number;
+  socketPath: string;
+  heartbeatAt: string;
+  ownerGeneration: number;
+  queueDepth: number;
+  alive: boolean;
+  stale: boolean;
+};
+
+function queueBaseDir(): string {
+  return path.join(os.homedir(), ".acpx", "queues");
+}
+
+function queueKeyForSession(sessionId: string): string {
+  return createHash("sha256").update(sessionId).digest("hex").slice(0, 24);
+}
+
+function queueLockFilePath(sessionId: string): string {
+  return path.join(queueBaseDir(), `${queueKeyForSession(sessionId)}.lock`);
+}
+
+function queueSocketPath(sessionId: string): string {
+  const key = queueKeyForSession(sessionId);
+  if (process.platform === "win32") {
+    return `\\\\.\\pipe\\acpx-${key}`;
+  }
+  return path.join(queueBaseDir(), `${key}.sock`);
+}
+
+function parseQueueOwnerRecord(raw: unknown): QueueOwnerRecord | null {
+  if (!raw || typeof raw !== "object" || Array.isArray(raw)) {
+    return null;
+  }
+  const record = raw as Record<string, unknown>;
+
+  if (
+    !Number.isInteger(record.pid) ||
+    (record.pid as number) <= 0 ||
+    typeof record.sessionId !== "string" ||
+    typeof record.socketPath !== "string" ||
+    typeof record.createdAt !== "string" ||
+    typeof record.heartbeatAt !== "string" ||
+    !Number.isInteger(record.ownerGeneration) ||
+    (record.ownerGeneration as number) <= 0 ||
+    !Number.isInteger(record.queueDepth) ||
+    (record.queueDepth as number) < 0
+  ) {
+    return null;
+  }
+
+  return {
+    pid: record.pid as number,
+    sessionId: record.sessionId,
+    socketPath: record.socketPath,
+    createdAt: record.createdAt,
+    heartbeatAt: record.heartbeatAt,
+    ownerGeneration: record.ownerGeneration as number,
+    queueDepth: record.queueDepth as number,
+  };
+}
+
+function createOwnerGeneration(): number {
+  return Date.now() * 1_000 + Math.floor(Math.random() * 1_000);
+}
+
+function nowIso(): string {
+  return new Date().toISOString();
+}
+
+function isQueueOwnerHeartbeatStale(owner: QueueOwnerRecord): boolean {
+  const heartbeatMs = Date.parse(owner.heartbeatAt);
+  if (!Number.isFinite(heartbeatMs)) {
+    return true;
+  }
+  return Date.now() - heartbeatMs > QUEUE_OWNER_STALE_HEARTBEAT_MS;
+}
+
+async function ensureQueueDir(): Promise<void> {
+  await fs.mkdir(queueBaseDir(), { recursive: true });
+}
+
+async function removeSocketFile(socketPath: string): Promise<void> {
+  if (process.platform === "win32") {
+    return;
+  }
+
+  try {
+    await fs.unlink(socketPath);
+  } catch (error) {
+    if ((error as NodeJS.ErrnoException).code !== "ENOENT") {
+      throw error;
+    }
+  }
+}
+
+async function waitForProcessExit(pid: number, timeoutMs: number): Promise<boolean> {
+  const deadline = Date.now() + Math.max(0, timeoutMs);
+  while (Date.now() <= deadline) {
+    if (!isProcessAlive(pid)) {
+      return true;
+    }
+    await waitMs(PROCESS_POLL_MS);
+  }
+
+  return !isProcessAlive(pid);
+}
+
+async function cleanupStaleQueueOwner(
+  sessionId: string,
+  owner: QueueOwnerRecord | undefined,
+): Promise<void> {
+  const lockPath = queueLockFilePath(sessionId);
+  const socketPath = owner?.socketPath ?? queueSocketPath(sessionId);
+
+  await removeSocketFile(socketPath).catch(() => {
+    // ignore stale socket cleanup failures
+  });
+
+  await fs.unlink(lockPath).catch((error) => {
+    if ((error as NodeJS.ErrnoException).code !== "ENOENT") {
+      throw error;
+    }
+  });
+}
+
+export async function readQueueOwnerRecord(
+  sessionId: string,
+): Promise<QueueOwnerRecord | undefined> {
+  const lockPath = queueLockFilePath(sessionId);
+  try {
+    const payload = await fs.readFile(lockPath, "utf8");
+    const parsed = parseQueueOwnerRecord(JSON.parse(payload));
+    return parsed ?? undefined;
+  } catch {
+    return undefined;
+  }
+}
+
+export function isProcessAlive(pid: number | undefined): boolean {
+  if (!pid || !Number.isInteger(pid) || pid <= 0 || pid === process.pid) {
+    return false;
+  }
+
+  try {
+    process.kill(pid, 0);
+    return true;
+  } catch {
+    return false;
+  }
+}
+
+export async function terminateProcess(pid: number): Promise<boolean> {
+  if (!isProcessAlive(pid)) {
+    return false;
+  }
+
+  try {
+    process.kill(pid, "SIGTERM");
+  } catch {
+    return false;
+  }
+
+  if (await waitForProcessExit(pid, PROCESS_EXIT_GRACE_MS)) {
+    return true;
+  }
+
+  try {
+    process.kill(pid, "SIGKILL");
+  } catch {
+    return false;
+  }
+
+  await waitForProcessExit(pid, PROCESS_EXIT_GRACE_MS);
+  return true;
+}
+
+export async function ensureOwnerIsUsable(
+  sessionId: string,
+  owner: QueueOwnerRecord,
+): Promise<boolean> {
+  const alive = isProcessAlive(owner.pid);
+  const stale = isQueueOwnerHeartbeatStale(owner);
+  if (alive && !stale) {
+    return true;
+  }
+
+  if (alive) {
+    await terminateProcess(owner.pid).catch(() => {
+      // best effort stale owner termination
+    });
+  }
+  await cleanupStaleQueueOwner(sessionId, owner);
+  return false;
+}
+
+export async function readQueueOwnerStatus(
+  sessionId: string,
+): Promise<QueueOwnerStatus | undefined> {
+  const owner = await readQueueOwnerRecord(sessionId);
+  if (!owner) {
+    return undefined;
+  }
+
+  const alive = await ensureOwnerIsUsable(sessionId, owner);
+  if (!alive) {
+    return undefined;
+  }
+
+  return {
+    pid: owner.pid,
+    socketPath: owner.socketPath,
+    heartbeatAt: owner.heartbeatAt,
+    ownerGeneration: owner.ownerGeneration,
+    queueDepth: owner.queueDepth,
+    alive,
+    stale: isQueueOwnerHeartbeatStale(owner),
+  };
+}
+
+export async function tryAcquireQueueOwnerLease(
+  sessionId: string,
+  nowIsoFactory: () => string = nowIso,
+): Promise<QueueOwnerLease | undefined> {
+  await ensureQueueDir();
+  const lockPath = queueLockFilePath(sessionId);
+  const socketPath = queueSocketPath(sessionId);
+  const createdAt = nowIsoFactory();
+  const ownerGeneration = createOwnerGeneration();
+  const payload = JSON.stringify(
+    {
+      pid: process.pid,
+      sessionId,
+      socketPath,
+      createdAt,
+      heartbeatAt: createdAt,
+      ownerGeneration,
+      queueDepth: 0,
+    },
+    null,
+    2,
+  );
+
+  try {
+    await fs.writeFile(lockPath, `${payload}\n`, {
+      encoding: "utf8",
+      flag: "wx",
+    });
+    await removeSocketFile(socketPath).catch(() => {
+      // best-effort stale socket cleanup after ownership is acquired
+    });
+    return {
+      sessionId,
+      lockPath,
+      socketPath,
+      createdAt,
+      ownerGeneration,
+    };
+  } catch (error) {
+    if ((error as NodeJS.ErrnoException).code !== "EEXIST") {
+      throw error;
+    }
+
+    const owner = await readQueueOwnerRecord(sessionId);
+    if (!owner) {
+      await cleanupStaleQueueOwner(sessionId, owner);
+      return undefined;
+    }
+
+    if (!isProcessAlive(owner.pid) || isQueueOwnerHeartbeatStale(owner)) {
+      if (isProcessAlive(owner.pid)) {
+        await terminateProcess(owner.pid).catch(() => {
+          // best effort stale owner termination
+        });
+      }
+      await cleanupStaleQueueOwner(sessionId, owner);
+    }
+    return undefined;
+  }
+}
+
+export async function refreshQueueOwnerLease(
+  lease: QueueOwnerLease,
+  options: {
+    queueDepth: number;
+  },
+  nowIsoFactory: () => string = nowIso,
+): Promise<void> {
+  const payload = JSON.stringify(
+    {
+      pid: process.pid,
+      sessionId: lease.sessionId,
+      socketPath: lease.socketPath,
+      createdAt: lease.createdAt,
+      heartbeatAt: nowIsoFactory(),
+      ownerGeneration: lease.ownerGeneration,
+      queueDepth: Math.max(0, Math.round(options.queueDepth)),
+    },
+    null,
+    2,
+  );
+  await fs.writeFile(lease.lockPath, `${payload}\n`, {
+    encoding: "utf8",
+  });
+}
+
+export async function releaseQueueOwnerLease(lease: QueueOwnerLease): Promise<void> {
+  await removeSocketFile(lease.socketPath).catch(() => {
+    // ignore best-effort cleanup failures
+  });
+
+  await fs.unlink(lease.lockPath).catch((error) => {
+    if ((error as NodeJS.ErrnoException).code !== "ENOENT") {
+      throw error;
+    }
+  });
+}
+
+export async function terminateQueueOwnerForSession(sessionId: string): Promise<void> {
+  const owner = await readQueueOwnerRecord(sessionId);
+  if (!owner) {
+    return;
+  }
+
+  if (isProcessAlive(owner.pid)) {
+    await terminateProcess(owner.pid);
+  }
+
+  await cleanupStaleQueueOwner(sessionId, owner);
+}
+
+export async function waitMs(ms: number): Promise<void> {
+  await new Promise<void>((resolve) => {
+    setTimeout(resolve, ms);
+  });
+}
